Refactor shared scraping and queue helpers
