+++ Shared.Tests/HtmlTextTests.cs
@@ -1,0 +1,17 @@
+using Shared.Services;
+
+namespace Shared.Tests;
+
+public class HtmlTextTests
+{
+    [Theory]
+    [InlineData(null, "")]
+    [InlineData("", "")]
+    [InlineData("  ", "")]
+    [InlineData("&lt;strong&gt;Ultra&lt;/strong&gt;", "<strong>Ultra</strong>")]
+    [InlineData("<p>Eco&nbsp;Trail</p><script>ignore()</script>", "Eco Trail ignore()")]
+    public void DecodeAndStripTags_ReturnsPlainDecodedText(string? html, string expected)
+    {
+        Assert.Equal(expected, HtmlText.DecodeAndStripTags(html));
+    }
+}
