+++ Shared/Services/RaceTypeNormalizer.cs
@@ -7,8 +7,9 @@ public static class RaceTypeNormalizer
         if (string.IsNullOrWhiteSpace(raceType))
             return null;
 
+        var separators = new[] { ',', ';', '/', '_' };
         var parts = raceType
-            .Split(new[] { ',', ';', '/', '_' }, StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
+            .Split(separators, StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries)
             .Select(p => p.ToLowerInvariant())
             .Select(p => RaceTypeAliases.TryGetValue(p, out var mapped) ? mapped : p)
             .Where(p => p.Length > 0)
